# [Enhancement][POC] optimize windows sort property when single tablet mode