@Seaven 2022-09-14
run starrocks_fe_unittest